+++ src/main/kotlin/icu/windea/pls/config/CwtConfigOption.kt
@@ -3,6 +3,7 @@ package icu.windea.pls.config
 data class CwtConfigOption(
 	val key:String,
 	val separator: CwtConfigSeparator = CwtConfigSeparator.EQUAL,
+	val value:String,
 	var booleanValue:Boolean? = null,
 	var intValue:Int? = null,
 	var floatValue:Float? = null,