+++ src/main/kotlin/icu/windea/pls/config/CwtConfigSeparator.kt
@@ -1,18 +1,19 @@
 package icu.windea.pls.config
 
-enum class CwtConfigSeparator(val key:String) {
-	EQUAL("="),NOT_EQUAL("<>");
+enum class CwtConfigSeparator(val key: String) {
+	EQUAL("="),
+	NOT_EQUAL("<>");
 	
 	override fun toString(): String {
 		return key
 	}
 	
-	companion object{
-		fun resolve(key:String): CwtConfigSeparator {
-			return when(key){
+	companion object {
+		fun resolve(key: String?): CwtConfigSeparator? {
+			return when(key) {
 				"=" -> EQUAL
-				"<>","!=" -> NOT_EQUAL
-				else -> EQUAL
+				"<>", "!=" -> NOT_EQUAL
+				else -> null
 			}
 		}
 	}