+++ src/main/kotlin/icu/windea/pls/model/ParadoxDefinitionLocalisationInfo.kt
@@ -1,8 +1,8 @@
 package icu.windea.pls.model
 
 data class ParadoxDefinitionLocalisationInfo(
-	val name:String,
-	val keyName:String,
-	val required:Boolean = false,
-	val primary:Boolean = false
+	val name: String,
+	val keyName: String,
+	val required: Boolean = false,
+	val primary: Boolean = false
 )